strategy

+++ Strategy/src/ICalcMedia.java
@@ -1,4 +1,4 @@
 public interface ICalcMedia {
-    public double CaculaMedia(int a, int b);
+    public double CaculaMedia(double a, double b);
     public String Situacao(double media);
 }
